requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiohttp-socks==0.7.1
aiosignal==1.3.1
altair==4.2.0
anyio==3.6.1
appdirs==1.4.4
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.5
cachetools==5.2.0
cchardet==2.1.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.5
commonmark==0.9.1
contourpy==1.0.5
cycler==0.11.0
dask==2022.10.0
debugpy==1.6.3
decorator==5.1.1
elastic-transport==8.4.0
elasticsearch==8.5.3
entrypoints==0.4
executing==1.1.1
fake-useragent==1.1.1
fastapi==0.85.0
fonttools==4.37.3
frozenlist==1.3.3
fsspec==2022.10.0
geographiclib==2.0
geopy==2.3.0
gitdb==4.0.10
GitPython==3.1.29
googletransx==2.4.2
h11==0.14.0
html5lib==1.1
htmlmin==0.1.12
httptools==0.5.0
idna==3.4
ImageHash==4.3.1
importlib-metadata==5.1.0
importlib-resources==5.10.2
inflection==0.5.1
install==1.3.5
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.3
jupyter_core==4.11.2
kiwisolver==1.4.4
locket==1.0.0
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mizani==0.8.1
more-itertools==9.0.0
multidict==6.0.3
multimethod==1.9.1
multitasking==0.0.11
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8
numpy==1.23.4
packaging==21.3
palettable==3.3.0
pandas==1.5.0
pandas-profiling==3.6.2
parso==0.8.3
partd==1.3.0
patsy==0.5.2
phik==0.12.3
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.11.0
plotly-express==0.4.1
plotnine==0.9.0
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.3
psycopg2-binary==2.9.4
pure-eval==0.2.2
pyarrow==10.0.1
pycares==4.3.0
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-socks==2.0.3
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
Quandl==3.7.0
regex==2022.10.31
requests==2.28.1
rich==12.6.0
schedule==1.1.0
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.1
semver==2.13.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.1
starlette==0.20.4
statsmodels==0.13.2
streamlit==1.15.2
streamlit-pandas-profiling==0.1.3
sweetviz==2.1.4
tangled-up-in-unicode==0.2.0
tenacity==8.1.0
textblob==0.17.1
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
twint==2.1.20
typeguard==2.13.3
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.18.3
validators==0.20.0
visions==0.7.5
watchdog==2.2.0
watchfiles==0.17.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
xgboost==1.7.3
yarl==1.8.2
yfinance==0.1.90
zipp==3.11.0
